requirements.txt
Django
gobiko.apns
